requirements.txt
jira
openpyxl
stringcase
PyYAML